申請取り下げを作成しました

+++ app/src/main/java/com/example/springboot/dto/input/WithDrowInput.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @NoArgsConstructor
 public class WithDrowInput
 {
-    private int requestId;
+    private Long requestId;
     private int requestType;
 }
